+++ funkode/core/constants.py
@@ -1,0 +1,5 @@
+"""Constants that are used all over FunKode."""
+import pathlib
+
+ROOT_DIR = pathlib.Path(__file__).resolve().parent.parent
+ICONS_DIR = ROOT_DIR/"icons"
