+++ funkode/core/palette.py
